requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.5
django-environ==0.4.4
Djangorestframework==3.7.3
gunicorn==19.7.1
idna==2.6
pytz==2017.2
requests==2.18.4
six==1.11.0
urllib3==1.22
whitenoise==3.3.1
